Scale catalog, tracker, and metadata upload state
Carry complete metainfo tracker tiers into durable and runtime state, retain unsupported HTTP(S) rows truthfully, redact presentation URLs, and remove the old u8/32-record assumptions.

Page file and tracker views at 1,024 rows, render catalogs lazily, and add compact all/none/range intake selection plus ranged priority mutation across generated web, gateway, desktop, and UniFFI contracts.

Match libtorrent's BEP 9 upload backpressure with a 160 KiB send watermark and a 1,024-request deferred occupancy bound while allowing more than 1,024 requests over a connection lifetime.

Topic: application-control

Topic: download-correctness

Topic: capability-readiness

+++ crates/rstorrent-engine/src/driver/control.rs
@@ -87,7 +87,7 @@ pub enum DownloadActivityEvent {
     StorageState(Box<DiskRuntimeSnapshot>),
     TrackerAnnounceStarted {
         tracker: String,
-        tier: u8,
+        tier: u32,
         attempt: u32,
         event: AnnounceEvent,
     },
@@ -103,7 +103,7 @@ pub enum DownloadActivityEvent {
     },
     TrackerFallbackSelected {
         tracker: String,
-        tier: u8,
+        tier: u32,
     },
     TrackerRetryScheduled {
         tracker: String,
